cleans up applicationHelper a bit
