+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers/Objects/ObjectFactory.cs
@@ -1,13 +1,15 @@
-﻿namespace Relativity.Test.Helpers.Objects
+﻿using Relativity.API;
+
+namespace Relativity.Test.Helpers.Objects
 {
 	public static class ObjectFactory
 	{
 
-		public static ObjectHelper ObjectHelperInstance()
+		public static ObjectHelper ObjectHelperInstance(IHelper helper)
 		{
 			var objectHelper = new ObjectHelper
 			{
-				Application = new Application.ApplicationHelper(),
+				Application = new Application.ApplicationHelper(helper),
 				Client = new Client.ClientHelper(),
 				Document = new Document.DocumentHelper(),
 				Folder = new Folder.FolderHelper(),
