+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers/Relativity.Test.Helpers.csproj
@@ -214,7 +214,7 @@
   </Choose>
   <ItemGroup>
     <Compile Include="Configuration\ConfigurationFactory.cs" />
-    <Compile Include="Objects\Application\ApplicationHelpers.cs" />
+    <Compile Include="Objects\Application\ApplicationHelper.cs" />
     <Compile Include="Objects\Application\ApplicationInstallException.cs" />
     <Compile Include="Objects\Client\ClientHelper.cs" />
     <Compile Include="Objects\Document\DocumentHelper.cs" />
